+++ src/pages/integration-builder/builder/commonSnippets/placeholders.ts
@@ -11,4 +11,6 @@ export const PLACEHOLDERS = {
   RN_RESOLVED_REDIRECT_URL: "getRNResolvedRedirectUrl",
   ANDROID_LOGIN_CONFIG: "getAndroidLoginConfig",
   FLUTTER_LOGIN_CONFIG: "getFlutterLoginConfig",
+  WEB3AUTH_OPTIONS: "Web3AuthOptions",
+  LOGIN_CONFIG: "LoginParams",
 };
