+++ src/pages/integration-builder/builder/frameworks/unity/filenames.ts
@@ -1,0 +1,6 @@
+export const FILENAME_WEB3AUTH = "frameworks/unity/Web3Auth.cs";
+export const FILENAME_MANIFEST = "frameworks/unity/manifest.json";
+
+export default function getFileNames(filenames) {
+  filenames.push(FILENAME_WEB3AUTH, FILENAME_MANIFEST);
+}
