--- src/pages/integration-builder/builder/unitySnippets/index.ts
@@ -1,3 +1,0 @@
-export * from "./get-web3auth-options";
-export * from "./get-login-code";
-export * from "./placeholders";
