--- src/pages/integration-builder/builder/unitySnippets/placeholders.ts
@@ -1,4 +1,0 @@
-export const PLACEHOLDERS = {
-  WEB3AUTH_OPTIONS: "Web3AuthOptions",
-  LOGIN_CONFIG: "LoginParams",
-};
